+++ src/app/src/config.cpp
@@ -6,8 +6,12 @@
 #include "config.h"
 #include <pm/Camera.h>
 
-Config::Config(std::filesystem::path cfg, cxxopts::ParseResult userargs) {
+Config::Config(std::filesystem::path cfg, std::filesystem::path profile, cxxopts::ParseResult userargs) {
     toml::value config;
+
+    //save user profile path
+    userProfile = profile;
+
     try {
         config = toml::parse<toml::preserve_comments, tsl::ordered_map>(cfg.string());
         configFile = cfg.string();
@@ -17,7 +21,7 @@ Config::Config(std::filesystem::path cfg, cxxopts::ParseResult userargs) {
     }
 
     try {
-        machineVarsFilePath = toml::find<std::string>(config, "nautilai", "machine_vars_file_path");
+        machineVarsFilePath = std::filesystem::path("C:\\ProgramData\\Curi Bio\\Nautilai\\machine.toml");
         machineVars = toml::parse<toml::preserve_comments, tsl::ordered_map>(machineVarsFilePath.string());
     } catch(const std::exception& e) {
         spdlog::error("Failed to parse machine vars file \"{}\"", e.what());
@@ -229,6 +233,9 @@ Config::Config(std::filesystem::path cfg, cxxopts::ParseResult userargs) {
        // exception occurs if machine or nautilai.toml files are missing
        spdlog::error("Unable to parse config variables due to missing toml file(s)");
     }
+
+    //updates default to false, will get set when manifest is downloaded
+    updateAvailable = false;
 
     //debug
     testImgPath = "";