+++ CHANGELOG.rst
@@ -6,6 +6,9 @@ Changelog for Nautilai
 Added:
 ^^^^^^
 - updated nautilai.toml file
+- autoupdate functionality
+- switched installer to per-user location
+- removed machine_vars_path from nautilai.toml and hard-coded to C:\ProgramData\Curi Bio\Nautilai\machine.toml
 
 Changed:
 ^^^^^^